Major Changes
-----------------------------------------------
Assets can now only be deleted by users who are the file owner or the admin.


git-svn-id: svn+ssh://svn.locusfoc.us/home/svn/ram/trunk@939 7bd17f4f-1217-0410-bbc5-bc4367116c30

+++ app/controllers/asset_controller.rb
@@ -141,6 +141,7 @@ class AssetController < ProtectedController
     raise if request.get?
     @asset = find_asset_by params[:id]
     raise if @asset.nil?
+    raise unless @asset.user_id == current_user.id || current_user.is_admin?
     @category = Category.find(@asset.category_id)
     raise unless accessible_items(@category,'assets',current_user.groups).include?(@asset)
     raise unless @asset.destroy
